move convertPeriodToDuration and getRegions into aws.go
