+++ x-pack/metricbeat/module/aws/aws.go
@@ -4,7 +4,17 @@
 
 package aws
 
-import "github.com/elastic/beats/metricbeat/mb"
+import (
+	"strconv"
+
+	awssdk "github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/aws/defaults"
+	"github.com/aws/aws-sdk-go-v2/service/ec2"
+	"github.com/aws/aws-sdk-go-v2/service/ec2/ec2iface"
+	"github.com/pkg/errors"
+
+	"github.com/elastic/beats/metricbeat/mb"
+)
 
 // Config defines all required and optional parameters for aws metricsets
 type Config struct {
@@ -18,6 +28,10 @@ type Config struct {
 // MetricSet is the base metricset for all aws metricsets
 type MetricSet struct {
 	mb.BaseMetricSet
+	RegionsList    []string
+	DurationString string
+	PeriodInSec    int
+	AwsConfig      *awssdk.Config
 }
 
 // ModuleName is the name of this module.
@@ -44,5 +58,79 @@ func NewMetricSet(base mb.BaseMetricSet) (*MetricSet, error) {
 	if err != nil {
 		return nil, err
 	}
-	return &MetricSet{BaseMetricSet: base}, nil
+
+	awsConfig := defaults.Config()
+	awsCreds := awssdk.Credentials{
+		AccessKeyID:     config.AccessKeyID,
+		SecretAccessKey: config.SecretAccessKey,
+	}
+	if config.SessionToken != "" {
+		awsCreds.SessionToken = config.SessionToken
+	}
+
+	awsConfig.Credentials = awssdk.StaticCredentialsProvider{
+		Value: awsCreds,
+	}
+
+	awsConfig.Region = config.DefaultRegion
+
+	svcEC2 := ec2.New(awsConfig)
+	regionsList, err := getRegions(svcEC2)
+	if err != nil {
+		return nil, err
+	}
+
+	// Calculate duration based on period
+	durationString, periodSec, err := convertPeriodToDuration(config.Period)
+	if err != nil {
+		return nil, err
+	}
+
+	// Construct MetricSet
+	metricSet := MetricSet{
+		BaseMetricSet:  base,
+		RegionsList:    regionsList,
+		DurationString: durationString,
+		PeriodInSec:    periodSec,
+		AwsConfig:      &awsConfig,
+	}
+	return &metricSet, nil
+}
+
+func getRegions(svc ec2iface.EC2API) (regionsList []string, err error) {
+	input := &ec2.DescribeRegionsInput{}
+	req := svc.DescribeRegionsRequest(input)
+	output, err := req.Send()
+	if err != nil {
+		err = errors.Wrap(err, "Failed DescribeRegions")
+		return
+	}
+	for _, region := range output.Regions {
+		regionsList = append(regionsList, *region.RegionName)
+	}
+	return
+}
+
+func convertPeriodToDuration(period string) (string, int, error) {
+	// Set starttime double the default frequency earlier than the endtime in order to make sure
+	// GetMetricDataRequest gets the latest data point for each metric.
+	numberPeriod, err := strconv.Atoi(period[0 : len(period)-1])
+	if err != nil {
+		return "", 0, err
+	}
+
+	unitPeriod := period[len(period)-1:]
+	switch unitPeriod {
+	case "s":
+		duration := "-" + strconv.Itoa(numberPeriod*2) + unitPeriod
+		return duration, numberPeriod, nil
+	case "m":
+		duration := "-" + strconv.Itoa(numberPeriod*2) + unitPeriod
+		periodInSec := numberPeriod * 60
+		return duration, periodInSec, nil
+	default:
+		err = errors.New("invalid period in config. Please reset period in config")
+		duration := "-" + strconv.Itoa(numberPeriod*2) + "s"
+		return duration, numberPeriod, err
+	}
 }
